drm: Move sysctl tree init/cleanup in drm_debugfs.c
Before, the sysctl tree was initialized from drm_sysfs.c. However, it
was never destroyed on unload and, unfortunately, there is no good place
to do it. In particular, it is too late in drm_sysfs_release() because
the device private data pointer (device_get_drvdata()) is already
cleared.

Therefore, the initialization and cleanup are made from drm_debugfs.c.