Various 4.11 fixes
- [drm]  Added drm/vc4 Raspberry Pi driver code and drm/vc4 module
         (still missing lkpi functionality for ARM platform so source files
		  commented out in Makefile)
- [i915] Added HuC firmware
- [lkpi] Fixed scatterlist so that GuC RSA data now uploads successfully
- [lkpi] Moved relay from common -> gplv2

Tested on Skylake:
HuC and DMC loads fine.
GuC now loads but because of some other reason the GPU hangs and after GPU reset
GuC firmware fails to load the second time. Reason is that the RSA data is not the
data from the firmware file but some random stuff (mostly zeros).
FB Helper is NULL (reason unknown).